Fix unsafe buffer usage in remoting/protocol
Replace UNSAFE_TODO macros, raw pointer arithmetic, and memcpy calls
with bounds-checked base::span and base::HeapArray operations. Updating
APIs like TcpWritePacket to accept spans instead of raw pointers and
sizes improves memory safety.

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7594097
Commit-Queue: kelsen liu <kelsenliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1589832}

+++ remoting/protocol/fake_stream_socket.h
@@ -6,6 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef REMOTING_PROTOCOL_FAKE_STREAM_SOCKET_H_
 #define REMOTING_PROTOCOL_FAKE_STREAM_SOCKET_H_
 
+#include <cstddef>
 #include <map>
 #include <memory>
 #include <optional>
@@ -96,13 +97,13 @@ class FakeStreamSocket : public P2PStreamSocket {
 
   std::optional<int> next_read_error_;
   scoped_refptr<net::IOBuffer> read_buffer_;
-  int read_buffer_size_ = 0;
+  size_t read_buffer_size_ = 0;
   net::CompletionOnceCallback read_callback_;
   base::WeakPtr<FakeStreamSocket> peer_socket_;
 
   std::string written_data_;
   std::string input_data_;
-  int input_pos_ = 0;
+  size_t input_pos_ = 0;
 
   scoped_refptr<base::SingleThreadTaskRunner> task_runner_;
   base::WeakPtrFactory<FakeStreamSocket> weak_factory_{this};
